WithFilters

+++ src/main/java/com/example/sweater/service/UserService.java
@@ -1,23 +1,32 @@
 package com.example.sweater.service;
 
+import com.example.sweater.Models.UserCriteriaRep;
+import com.example.sweater.Models.UserPage;
+import com.example.sweater.Models.UserSearchCriteria;
 import com.example.sweater.Models.UserUpdateRequesModel;
 import com.example.sweater.entity.User;
 import com.example.sweater.repos.UserRep;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
     @Autowired
     UserRep userRep;
+    @Autowired
+    private UserCriteriaRep userCriteriaRep;
 
-    public void save(User user){
-        userRep.save(user);
+    public User save(User user){
+       User newUser= userRep.save(user);
+       return newUser;
     }
     public List<User> getUsers(){
+
         return userRep.findAll();
     }
 
@@ -32,7 +41,17 @@ public class UserService {
     return user;
     }
 
-    public void delete(Long id) {
-        userRep.deleteById(id);
+    public User getUserFromId(Long id){
+        return userRep.findById(id).orElse(null);
+    }
+    public User delete(Long id) {
+        User user=userRep.findById(id).orElse(null);
+        if(user!=null)  userRep.deleteById(id);
+        return user;
+
+    }
+
+    public List<User> getUsers(UserPage userPage, UserSearchCriteria userSearchCriteria){
+        return userCriteriaRep.FindAllWithFilters(userPage,userSearchCriteria);
     }
 }
